chore: integerate changes to ptoFusionPlan pass
