+++ lib/PTO/Transforms/CMakeLists.txt
@@ -17,7 +17,6 @@ add_mlir_dialect_library(PTOTransforms
   TileFusion/PTOPreFusionAnalysis.cpp
   TileFusion/PTOPrintPreFusionAnalysis.cpp
   TileFusion/PTOFusionPlan.cpp
-  TileFusion/PTOVersionAwareFusionPlan.cpp
   TileFusion/PTOOpScheduling.cpp
   TileFusion/PTOMarkLastUse.cpp
   TileFusion/PTOFusionRegionGen.cpp
